ENH: Refactored the training problem

+++ source/models/ClassTransformerModel.py
@@ -496,7 +496,9 @@ class ClassTransformerModel:
             assignmentLosses.append(self.evaluatePerAssignmentClusterLoss(
                 features[:, :, :, i, :], classLabels[:, :, :, i]))
 
-        return sum(assignmentLosses) / (tf.multiply(batchSize, tf.multiply(sequenceLength, 2 * self.getAssignmentCount())))
+        return sum(assignmentLosses) / (tf.multiply(tf.cast(batchSize, dtype=tf.float32),
+            tf.multiply(tf.cast(sequenceLength, dtype=tf.float32),
+            2.0 * self.getAssignmentCount())))
 
     def evaluatePerAssignmentClusterLoss(self, features, labels):
         # features is [batch, sequence, 2, feature-dim]
